Moved InProgressChapterFileFormat to project package
